pyproject.toml
# This is the modern, universal format for pyproject.toml

[project]
name = "adaptkan"
version = "0.1.0"
description = "Adaptable Kolmogorov-Arnold Networks with support for continual learning in PyTorch and JAX."
readme = "README.md"
license = { text = "GPL-3.0" }
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10"
dependencies = [
    "Cython==3.1.3",
    "cvxpy",
    "numpy==1.26.0",
    "absl-py==2.3.1",
    "beautifulsoup4==4.13.4",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "chex==0.1.86",
    "colorlog==6.9.0",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "diffdist==0.1",
    "eagerpy==0.30.0",
    "equinox==0.12.1",
    "faiss-cpu==1.7.2",
    "filelock==3.19.1",
    "fonttools==4.59.1",
    "foolbox==3.3.4",
    "fsspec==2025.7.0",
    "gdown==5.2.0",
    "GitPython==3.1.45",
    "greenlet==3.2.4",
    "hf-xet==1.1.7",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "imageio==2.37.0",
    "imgaug==0.4.0",
    "jaxlib==0.4.38",
    "jaxtyping==0.3.2",
    "Jinja2==3.1.6",
    "joblib==1.5.1",
    "json5==0.12.1",
    "kiwisolver==1.4.9",
    "lazy_loader==0.4",
    "lightning>=2.4.0",
    "Mako==1.3.10",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.5",
    "ml_dtypes",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "opencv-python==4.8.1.78",
    "opt-einsum==3.3.0",
    "optax==0.2.3",
    "optuna==4.5.0",
    "packaging==24.1",
    "pandas==2.3.1",
    "patsy==1.0.1",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "pykan==0.2.8",
    "pyparsing==3.2.3",
    "PySocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "requests==2.32.4",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.3.2",
    "scipy==1.11.4",
    "shapely==2.1.1",
    "six==1.17.0",
    "SQLAlchemy==2.0.43",
    "statsmodels==0.14.5",
    "sympy==1.13.1",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "timm==1.0.19",
    "tomli==2.2.1",
    "toolz==1.0.0",
    "typeguard==4.4.4",
    "typing_extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wadler_lindig==0.1.7",
    "seaborn",
    "cvxpylayers",
    "gurobipy",
    "qpax",
    "tensorboard",
    "pykan",
    "jupyter",
    "plotly"
    # Note: nvidia-* packages are often implicitly installed by jax/pytorch with CUDA
]

[project.optional-dependencies]
cuda = [
    "libmr==0.1.9",
    "jax[cuda12_pip]==0.4.38",
    "faiss-gpu==1.7.2",
    "openood @ git+https://github.com/Jingkang50/OpenOOD.git@a485cbb49d692b5f88e38b5ae2b68331f277b9ea",
    "triton==3.1.0"
    ]
cpu = [
    "jax[cpu]==0.4.38",
    "faiss-cpu==1.7.2"
    ]

# --- Build System Configuration ---
# This tells tools HOW to build your package. We use the standard setuptools.
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- Setuptools Configuration ---
# This tells setuptools WHERE to find your code.
[tool.setuptools]
packages = ["adaptkan"]

# Add this to your pyproject.toml
[tool.uv.extra-build-dependencies]
libmr = ["numpy", "cython"]